fix: extract output config from preamble
Setting formats in preambles would get confusing when merging files.
